+++ app/middleware/jwt.go
@@ -10,7 +10,7 @@ import (
 	"time"
 )
 
-func JWTAuth(GuardName string) gin.HandlerFunc {
+func JWTAuth(guardName string) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		tokenStr := c.Request.Header.Get("Authorization")
 		if tokenStr == "" {
@@ -32,7 +32,7 @@ func JWTAuth(GuardName string) gin.HandlerFunc {
 
 		claims := token.Claims.(*services.CustomClaims)
 		// Token 发布者校验
-		if claims.Issuer != GuardName {
+		if claims.Issuer != guardName {
 			response.TokenFail(c)
 			c.Abort()
 			return
@@ -42,12 +42,12 @@ func JWTAuth(GuardName string) gin.HandlerFunc {
 		if claims.ExpiresAt.Time.Unix()-time.Now().Unix() < global.App.Config.Jwt.RefreshGracePeriod {
 			lock := global.Lock("refresh_token_lock", global.App.Config.Jwt.JwtBlacklistGracePeriod)
 			if lock.Get() {
-				err, user := services.JwtService.GetUserInfo(GuardName, claims.ID)
+				user, err := services.JwtService.GetUserInfo(guardName, claims.ID)
 				if err != nil {
 					global.App.Log.Error(err.Error())
 					lock.Release()
 				} else {
-					tokenData, _, _ := services.JwtService.CreateToken(GuardName, user)
+					tokenData, _, _ := services.JwtService.CreateToken(guardName, user)
 					c.Header("new-token", tokenData.AccessToken)
 					c.Header("new-expires-in", strconv.Itoa(tokenData.ExpiresIn))
 					_ = services.JwtService.JoinBlackList(token)